Moved hparams + checkpoint test to test_hparams
